refactor(locale): rename ge to ka_GE
