+++ src/locale/ge.ts
@@ -1,13 +1,12 @@
-/*
- * This file is automatically generated.
- * Run 'pnpm run generate:locales' to update.
- */
+import { deprecated } from '../internal/deprecated';
+import { faker as ka_GE } from './ka_GE';
 
-import { Faker } from '../faker';
-import base from '../locales/base';
-import en from '../locales/en';
-import ge from '../locales/ge';
-
-export const faker = new Faker({
-  locale: [ge, en, base],
-});
+export const faker = (() => {
+  deprecated({
+    deprecated: "import { faker } from '@faker-js/faker/locale/ge'",
+    proposed: "import { faker } from '@faker-js/faker/locale/ka_GE'",
+    since: '8.0',
+    until: '9.0',
+  });
+  return ka_GE;
+})();
