+++ src/locale/ka_GE.ts
@@ -1,0 +1,13 @@
+/*
+ * This file is automatically generated.
+ * Run 'pnpm run generate:locales' to update.
+ */
+
+import { Faker } from '../faker';
+import base from '../locales/base';
+import en from '../locales/en';
+import ka_GE from '../locales/ka_GE';
+
+export const faker = new Faker({
+  locale: [ka_GE, en, base],
+});
