+++ src/locales/index.ts
@@ -34,7 +34,6 @@ export { default as fr_BE } from './fr_BE';
 export { default as fr_CA } from './fr_CA';
 export { default as fr_CH } from './fr_CH';
 export { default as fr_LU } from './fr_LU';
-export { default as ge } from './ge';
 export { default as he } from './he';
 export { default as hr } from './hr';
 export { default as hu } from './hu';
@@ -42,6 +41,7 @@ export { default as hy } from './hy';
 export { default as id_ID } from './id_ID';
 export { default as it } from './it';
 export { default as ja } from './ja';
+export { default as ka_GE } from './ka_GE';
 export { default as ko } from './ko';
 export { default as lv } from './lv';
 export { default as mk } from './mk';
