+++ src/locales/ka_GE/cell_phone/formats.ts
