+++ src/locales/ka_GE/cell_phone/index.ts
