+++ src/locales/ka_GE/company/index.ts
