+++ src/locales/ka_GE/company/name_patterns.ts
