+++ src/locales/ka_GE/internet/domain_suffix.ts
