+++ src/locales/ka_GE/internet/index.ts
