+++ src/locales/ka_GE/location/building_number.ts
