+++ src/locales/ka_GE/location/city.ts
