+++ src/locales/ka_GE/location/city_name.ts
