+++ src/locales/ka_GE/location/city_suffix.ts
