+++ src/locales/ka_GE/location/default_country.ts
