+++ src/locales/ka_GE/location/index.ts
