+++ src/locales/ka_GE/location/street_name.ts
