+++ src/locales/ka_GE/metadata.ts
@@ -1,7 +1,7 @@
 import type { MetadataDefinitions } from '../..';
 
 const metadata: MetadataDefinitions = {
-  title: 'Georgian',
+  title: 'Georgian (Georgia)',
 };
 
 export default metadata;
