+++ src/locales/ka_GE/person/female_prefix.ts
