+++ src/locales/ka_GE/person/first_name.ts
