+++ src/locales/ka_GE/person/index.ts
