+++ src/locales/ka_GE/person/last_name.ts
