+++ src/locales/ka_GE/person/last_name_patterns.ts
