+++ src/locales/ka_GE/person/male_prefix.ts
