+++ src/locales/ka_GE/person/name.ts
