+++ src/locales/ka_GE/phone_number/index.ts
